Refactor models to improve code
