+++ Tegami/Source/Models/AnimationConfig.swift
