+++ Tegami/Source/Models/FilmPosition.swift
@@ -1,0 +1,12 @@
+//
+//  FilmList.swift
+//  GhibliAPP
+//
+//  Created by Yago Marques on 19/09/22.
+//
+
+import Foundation
+
+struct FilmPosition: Codable, Equatable {
+    let filmId: String
+}
