+++ Tegami/Source/Models/GenreInfo.swift
@@ -1,0 +1,13 @@
+//
+//  GenreInfo.swift
+//  Tegami
+//
+//  Created by user on 11/10/22.
+//
+
+import Foundation
+
+struct GenreInfo: Decodable {
+    let name: String
+    let id: Int
+}
