+++ Tegami/Source/Models/GhibliInfo.swift
@@ -1,0 +1,15 @@
+//
+//  GhibliInfo.swift
+//  Tegami
+//
+//  Created by user on 11/10/22.
+//
+
+import Foundation
+
+struct GhibliInfo: Codable, Equatable {
+    let id: String
+    let releaseDate: String
+    let runningTime: String
+    let originalTitle: String
+}
