+++ Tegami/Source/Models/TmdbInfo.swift
@@ -1,0 +1,12 @@
+//
+//  TmdbInfo.swift
+//  Tegami
+//
+//  Created by user on 11/10/22.
+//
+
+import Foundation
+
+struct TmdbInfo: Codable {
+    let results: [TmdbResult]
+}
